+++ partition.go
@@ -4,7 +4,7 @@ import (
 	"io"
 
 	errors "gopkg.in/src-d/go-errors.v1"
-	"gopkg.in/src-d/go-mysql-server.v0/sql"
+	"github.com/src-d/go-mysql-server/sql"
 )
 
 // partitioned is an embeddable helper that contains the methods for a table