+++ utils.go
@@ -3,7 +3,7 @@ package gitbase
 import (
 	"io"
 
-	"gopkg.in/src-d/go-mysql-server.v0/sql"
+	"github.com/src-d/go-mysql-server/sql"
 )
 
 var noRows emptyRowIter